Revert "Aggiunta bluetooth"
This reverts commit d19b07ebb61f52673a63d78754d630380966024b.

+++ robot2/input_data.h
@@ -6,6 +6,6 @@
 bool check_obstacle(float dist_obs);
 
 //restituisce il percoso leto dai linefollower
-uint8_t read_path(uint8_t sensorStateDX, uint8_t sensorStateSX);
+int read_path(int sensorStateDX, int sensorStateSX);
 
 #endif
